requirements.txt
customtkinter
pillow
tensorflow
ctkmessagebox